+++ lib/src/widget/labels_view.dart
@@ -22,6 +22,10 @@ class LabelsView extends StatelessWidget {
     return AnimatedBuilder(
       animation: controller,
       builder: (context, _) {
+        if (!controller.canLayout) {
+          return const SizedBox();
+        }
+
         final visibleNodes = controller.getVisibleNodes();
         final layout = controller.layout;